Put callback in smac directory to allow importing it in custom repos
